@@ -14,79 +14,72 @@
     "node": ">=6.0"
   },
   "dependencies": {
-    "babel-core": "^6.14.0",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "body-parser": "^1.15.2",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.0",
+    "babel-preset-react": "^6.23.0",
+    "body-parser": "^1.17.1",
     "compression": "^1.6.2",
-    "debug": "^2.2.0",
-    "express": "^4.14.0",
-    "express-session": "^1.14.1",
-    "express-socket.io-session": "^1.3.1",
-    "gulp-imagemin": "^3.0.3",
-    "imagemin-gifsicle": "^5.1.0",
-    "imagemin-jpegtran": "^5.0.2",
-    "imagemin-optipng": "^5.2.1",
-    "imagemin-svgo": "^5.2.0",
-    "jest-cli": "^15.1.1",
-    "jquery": "^3.1.1",
-    "lodash": "^4.15.0",
-    "mongo": "^0.1.0",
-    "mongoose": "^4.6.0",
-    "morgan": "^1.7.0",
-    "nodemailer": "^2.6.0",
-    "nodemailer-mailgun-transport": "^1.2.2",
+    "debug": "^2.6.3",
+    "express": "^4.15.2",
+    "express-session": "^1.15.1",
+    "express-socket.io-session": "^1.3.2",
+    "jest-cli": "^19.0.2",
+    "jquery": "^3.2.1",
+    "lodash": "^4.17.4",
+    "mongoose": "^4.9.1",
+    "morgan": "^1.8.1",
+    "nodemailer": "^3.1.8",
+    "nodemailer-mailgun-transport": "^1.2.4",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^4.0.0",
-    "pug": "^2.0.0-beta6",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-redux": "^4.4.5",
+    "pug": "^2.0.0-beta11",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.3",
     "redux": "^3.6.0",
-    "redux-devtools": "^3.3.1",
-    "semantic-ui-checkbox": "^2.2.3",
-    "semantic-ui-dimmer": "^2.2.3",
-    "semantic-ui-dropdown": "^2.2.3",
-    "semantic-ui-form": "^2.2.3",
-    "semantic-ui-modal": "^2.2.3",
-    "semantic-ui-popup": "^2.2.3",
-    "semantic-ui-progress": "^2.2.3",
-    "semantic-ui-transition": "^2.2.3",
-    "semantic-ui-visibility": "^2.2.3",
-    "serve-favicon": "^2.3.0",
-    "socket.io": "^1.4.8",
-    "socket.io-client": "^1.4.8"
+    "redux-devtools": "^3.3.2",
+    "semantic-ui-checkbox": "^2.2.9",
+    "semantic-ui-dimmer": "^2.2.9",
+    "semantic-ui-dropdown": "^2.2.9",
+    "semantic-ui-form": "^2.2.9",
+    "semantic-ui-modal": "^2.2.9",
+    "semantic-ui-popup": "^2.2.9",
+    "semantic-ui-progress": "^2.2.9",
+    "semantic-ui-transition": "^2.2.9",
+    "semantic-ui-visibility": "^2.2.9",
+    "serve-favicon": "^2.4.2",
+    "socket.io": "^1.7.3",
+    "socket.io-client": "^1.7.3"
   },
   "devDependencies": {
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-stage-1": "^6.13.0",
+    "babel-preset-stage-1": "^6.24.1",
     "babelify": "^7.3.0",
-    "browserify": "^13.1.0",
-    "eslint": "^3.9.1",
-    "eslint-config-google": "^0.6.0",
-    "eslint-plugin-import": "^1.15.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.2.2",
+    "browserify": "^14.3.0",
+    "eslint": "^3.19.0",
+    "eslint-config-google": "^0.7.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.10.3",
     "gulp": "^3.9.1",
-    "gulp-clean-css": "^2.0.12",
+    "gulp-clean-css": "^3.0.4",
     "gulp-eslint": "^3.0.1",
-    "gulp-imagemin": "^3.0.3",
+    "gulp-imagemin": "^3.2.0",
     "gulp-livereload": "^3.8.1",
     "gulp-plumber": "^1.1.0",
     "gulp-rename": "^1.2.2",
-    "gulp-sass": "^2.3.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-uglify": "^2.0.0",
+    "gulp-sass": "^3.1.0",
+    "gulp-sourcemaps": "^2.6.0",
+    "gulp-uglify": "^2.1.2",
     "gulp-wait": "0.0.2",
-    "node-notifier": "^4.6.1",
-    "rc-slider": "^5.1.3",
-    "react-test-renderer": "^15.3.1",
+    "node-notifier": "^5.1.2",
+    "rc-slider": "5.1.3",
+    "react-test-renderer": "^15.5.4",
     "require-dot-file": "^0.4.0",
-    "through2": "^2.0.1",
+    "through2": "^2.0.3",
     "vinyl-transform": "^1.0.0",
-    "yamljs": "^0.2.8"
+    "yamljs": "^0.2.9"
   },
   "scripts": {
     "test": "jest",
