@@ -13,12 +13,23 @@ export default class Gamechat extends React.Component {
 		this.handleClickedLeaveGame = this.handleClickedLeaveGame.bind(this);
 		this.handleClickedClaimButton = this.handleClickedClaimButton.bind(this);
 		this.handleClaimButtonClick = this.handleClaimButtonClick.bind(this);
+		this.handleWhitelistPlayers = this.handleWhitelistPlayers.bind(this);
+
 		this.state = {
 			chatFilter: 'All',
 			lock: false,
 			inputValue: '',
-			claim: ''
+			claim: '',
+			playersToWhitelist: []
 		};
+	}
+
+	handleWhitelistPlayers() {
+		this.setState({
+			playersToWhitelist: this.props.userList.list.filter(user => user.userName !== this.props.userInfo.userName).map(user => ({userName: user.userName, isSelected: true}))
+		});
+
+		$(this.whitelistModal).modal('show');
 	}
 
 	handleClickedLeaveGame() {
@@ -42,8 +53,9 @@ export default class Gamechat extends React.Component {
 		const {userInfo, gameInfo} = this.props;
 		this.scrollChats();
 
-		if (prevProps && userInfo.userName && prevProps.gameInfo.publicPlayersState.filter(player => player.isDead).length !== gameInfo.publicPlayersState.filter(player => player.isDead).length && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).isDead ||
-			(prevProps && userInfo.userName && gameInfo.gameState.phase === 'presidentSelectingPolicy' && (gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName) && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).governmentStatus === 'isPresident' || gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName) && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).governmentStatus === 'isChancellor') && prevProps.gameInfo.gameState.phase !== 'presidentSelectingPolicy')) {
+		if (prevProps && userInfo.userName && prevProps.gameInfo.publicPlayersState.filter(player => player.isDead).length !== gameInfo.publicPlayersState.filter(player => player.isDead).length && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).isDead
+			|| (prevProps && userInfo.userName && gameInfo.gameState.phase === 'presidentSelectingPolicy' && (gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName) && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).governmentStatus === 'isPresident'
+			|| gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName) && gameInfo.publicPlayersState.find(player => userInfo.userName === player.userName).governmentStatus === 'isChancellor') && prevProps.gameInfo.gameState.phase !== 'presidentSelectingPolicy')) {
 			this.setState({inputValue: ''});
 			$(this.gameChatInput).blur();
 		}
@@ -163,7 +175,23 @@ export default class Gamechat extends React.Component {
 	}
 
 	render() {
-		const {userInfo, gameInfo} = this.props;
+		const {userInfo, gameInfo} = this.props,
+			selectedWhitelistplayer = (playerName) => {
+				const {playersToWhitelist} = this.state,
+					playerIndex = playersToWhitelist.findIndex(player => player.userName === playerName);
+
+				playersToWhitelist[playerIndex].isSelected = !playersToWhitelist[playerIndex].isSelected;
+
+				this.setState(playersToWhitelist);
+			},
+			submitWhitelist = () => {
+				const whitelistPlayers = this.state.playersToWhitelist.filter(player => player.isSelected).map(player => player.userName);
+				this.props.socket.emit('updateGameWhitelist', {
+					uid: gameInfo.general.uid,
+					whitelistPlayers
+				});
+				$(this.whitelistModal).modal('hide');
+			};
 
 		return (
 			<section className="gamechat">
@@ -172,6 +200,11 @@ export default class Gamechat extends React.Component {
 					<a className={this.state.chatFilter === 'Chat' ? 'item active' : 'item'} onClick={this.handleChatFilterClick}>Chat</a>
 					<a className={this.state.chatFilter === 'Game' ? 'item active' : 'item'} onClick={this.handleChatFilterClick}>Game</a>
 					<i className={this.state.lock ? 'large lock icon' : 'large unlock alternate icon'} onClick={this.handleChatLockClick} />
+					{(() => {
+						if (userInfo.isSeated && gameInfo.general.private && !gameInfo.gameState.isStarted) {
+							return <div className='ui button whitelist' onClick={this.handleWhitelistPlayers}>Whitelist Players</div>;
+						}
+					})()}
 					<div className={
 						(() => {
 							let classes = 'ui primary button';
@@ -183,15 +216,6 @@ export default class Gamechat extends React.Component {
 							return classes;
 						})()
 					} onClick={this.handleClickedLeaveGame}>Leave Game</div>
-					<div className="ui basic fullscreen modal leavegamemodals" ref={c => {
-						this.leaveGameModal = c;
-					}}>
-						<h2 className="ui header">DANGER.  Leaving an in-progress game will ruin it for the other players (unless you've been executed).  Do this only in the case of a game already ruined by an AFK/disconnected player or if someone has already left.</h2>
-						<div className="ui green positive inverted leave-game button">
-							<i className="checkmark icon"></i>
-							Leave anyways
-						</div>
-					</div>
 				</section>
 				<section className={
 					(() => {
@@ -273,17 +297,24 @@ export default class Gamechat extends React.Component {
 								const {gameState, publicPlayersState} = this.props.gameInfo,
 									{userName} = this.props.userInfo,
 									isDead = (() => {
-										if (this.props.userInfo.isSeated && publicPlayersState.length && publicPlayersState.find(player => userName === player.userName)) {
+										if (this.props.userInfo.isSeated
+											&& publicPlayersState.length
+											&& publicPlayersState.find(player => userName === player.userName)) {
 											return publicPlayersState.find(player => userName === player.userName).isDead;
 										}
 									})(),
 									isGovernmentDuringPolicySelection = (() => {
-										if ((gameState.phase === 'presidentSelectingPolicy' || gameState.phase === 'chancellorSelectingPolicy') && userName) {
+										if ((gameState.phase === 'presidentSelectingPolicy'
+											|| gameState.phase === 'chancellorSelectingPolicy') && userName) {
 											return publicPlayersState.find(player => player.userName === userName).governmentStatus === 'isPresident' || publicPlayersState.find(player => player.userName === userName).governmentStatus === 'isChancellor';
 										}
 									})();
 
-								if (!this.props.userInfo.userName || (gameState.cardsDealt && !gameState.isDay) || (isDead && !gameState.isCompleted) || isGovernmentDuringPolicySelection || this.props.gameInfo.general.disableChat) {
+								if (!this.props.userInfo.userName
+									|| (isDead && !gameState.isCompleted)
+									|| isGovernmentDuringPolicySelection
+									|| this.props.gameInfo.general.disableChat
+									|| (this.props.gameInfo.general.private && !this.props.userInfo.isSeated)) {
 									classes += ' disabled';
 								}
 
@@ -302,6 +333,35 @@ export default class Gamechat extends React.Component {
 						}
 					})()}
 				</form>
+				<div className="ui basic fullscreen modal leavegamemodals" ref={c => {
+					this.leaveGameModal = c;
+				}}>
+					<h2 className="ui header">DANGER.  Leaving an in-progress game will ruin it for the other players (unless you've been executed).  Do this only in the case of a game already ruined by an AFK/disconnected player or if someone has already left.</h2>
+					<div className="ui green positive inverted leave-game button">
+						<i className="checkmark icon"></i>
+						Leave anyways
+					</div>
+				</div>
+				<div className="ui basic fullscreen modal whitelistmodal" ref={c => {
+					this.whitelistModal = c;
+				}}>
+					<h2 className="ui header">Select player(s) below to whitelist for seating in this private game.</h2>
+					<ul>
+						{this.state.playersToWhitelist.map((player, index) => {
+							const uid = Math.random().toString(36).substring(2);
+
+							return (
+								<li key={index}>
+									<input type="checkbox" id={uid} defaultChecked={true} onChange={() => {
+										selectedWhitelistplayer(player.userName);
+									}}/>
+									<label htmlFor={uid}>{player.userName}</label>
+								</li>
+							);
+						})}
+					</ul>
+					<div className="ui green positive inverted whitelist-submit button" onClick={submitWhitelist}>Submit</div>
+				</div>
 			</section>
 		);
 	}
@@ -316,5 +376,6 @@ Gamechat.propTypes = {
 	selectedPlayer: React.PropTypes.object,
 	userInfo: React.PropTypes.object,
 	gameInfo: React.PropTypes.object,
-	socket: React.PropTypes.object
+	socket: React.PropTypes.object,
+	userList: React.PropTypes.object
 };
