@@ -57,9 +57,12 @@ const {games, userList, generalChats} = require('./models'),
 				const {gameState, publicPlayersState} = game,
 					playerIndex = publicPlayersState.findIndex(player => player.userName === passport.user);
 
-				if (gameState.isStarted && !gameState.isCompleted) {
+				if (gameState.isTracksFlipped && !gameState.isCompleted) {
 					publicPlayersState[playerIndex].connected = false;
 					sendInProgressGameUpdate(game);
+				} else if (gameState.isStarted && !gameState.isCompleted) {
+					publicPlayersState[playerIndex].connected = false;
+					io.in(game.uid).emit('gameUpdate', game);
 				} else if (gameState.isCompleted && game.publicPlayersState.filter(player => !player.connected || player.leftGame).length === game.general.playerCount - 1) {
 					saveGame(game);
 					games.splice(games.indexOf(game), 1);
@@ -79,10 +82,18 @@ const {games, userList, generalChats} = require('./models'),
 		sendUserList();
 	};
 
-module.exports.updateSeatedUser = data => {
+module.exports.updateSeatedUser = (socket, data) => {
 	const game = games.find(el => el.general.uid === data.uid);
 
-	if (game) {
+	// prevents race condition between 1) taking a seat and 2) the game starting
+	if (game && game.gameState.isTracksFlipped) {
+		console.warn('player joined too late');
+		return;
+	}
+
+	if (game
+    && game.publicPlayersState.length < game.general.maxPlayersCount
+    && (!game.general.private || (game.general.private && data.password === game.private.privatePassword || game.general.private && game.general.whitelistedPlayers.includes(data.userName)))) {
 		const {publicPlayersState} = game;
 
 		publicPlayersState.push({
@@ -96,6 +107,8 @@ module.exports.updateSeatedUser = data => {
 				cardBack: {}
 			}
 		});
+
+		socket.emit('updateSeatForUser', true);
 
 		if (publicPlayersState.length === game.general.maxPlayersCount && !game.gameState.isStarted) { // sloppy but not trivial to get around
 			game.gameState.isStarted = true;
@@ -131,6 +144,11 @@ module.exports.handleAddNewGame = (socket, data) => {
 		unSeatedGameChats: []
 	};
 
+	if (data.general.private) {
+		data.private.privatePassword = data.general.private;
+		data.general.private = true;
+	}
+
 	games.push(data);
 	sendGameList();
 	socket.join(data.general.uid);
@@ -147,6 +165,13 @@ module.exports.handleAddNewGameChat = data => {
 	} else {
 		io.in(data.uid).emit('gameUpdate', secureGame(game));
 	}
+};
+
+module.exports.handleUpdateWhitelist = data => {
+	const game = games.find(el => el.general.uid === data.uid);
+
+	game.general.whitelistedPlayers = data.whitelistPlayers;
+	io.in(data.uid).emit('gameUpdate', secureGame(game));
 };
 
 module.exports.handleNewGeneralChat = data => {
@@ -192,7 +217,7 @@ module.exports.handleUserLeaveGame = (socket, data) => {
 	}
 
 	if (game && game.gameState.isStarted && data.isSeated) {
-		const playerIndex = game.private.seatedPlayers.findIndex(player => player.userName === data.userName);
+		const playerIndex = game.publicPlayersState.findIndex(player => player.userName === data.userName);
 
 		game.publicPlayersState[playerIndex].leftGame = true;
 
@@ -207,12 +232,13 @@ module.exports.handleUserLeaveGame = (socket, data) => {
 
 	if (game && data.isSeated && !game.gameState.isStarted) {
 		game.publicPlayersState.splice(game.publicPlayersState.findIndex(player => player.userName === data.userName), 1);
+		io.sockets.in(data.uid).emit('gameUpdate', game);
 	}
 
 	if (game && !game.publicPlayersState.length) {
 		io.sockets.in(data.uid).emit('gameUpdate', {});
 		games.splice(games.indexOf(game), 1);
-	} else if (game && game.gameState.isStarted) {
+	} else if (game && game.isTracksFlipped) {
 		sendInProgressGameUpdate(game);
 	}
 
